+++ tests/TestCase.php
@@ -11,7 +11,7 @@ class TestCase extends \PHPUnit\Framework\TestCase
     public function setUp(): void
     {
         parent::setUp();
-        Lang::setToDefaults();
+        Lang::resetToDefaults();
     }
 
     /**