+++ CHANGELOG.md
@@ -10,6 +10,9 @@
     - Added support for Chinese language 🇨🇳
     - Added support for Japanese language 🇯🇵
     - Rewrite the codebase to make it more extendable
+    - Added `setToDefaults` method to `Lang` class to reset overwrites and the language to default
+    - Added `current` method to `Lang` class to get the current language
+    - Added `getOverwrites` and `setOverwrites` methods to `Lang` class to get and set overwrites
 
 ## v3.0.9 (2024-11-01)
 - Typo in `README.md` file