+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     - Added `setToDefaults` method to `Lang` class to reset overwrites and the language to default
     - Added `current` method to `Lang` class to get the current language
     - Added `getOverwrites` and `setOverwrites` methods to `Lang` class to get and set overwrites
+    - Increased number limit to quadrillions. This is the highest number supported now `999,999,999,999,999,999`
 
 ## v3.0.9 (2024-11-01)
 - Typo in `README.md` file